Finsh lines architecture
