--- app/dashboard/lines/[id]/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-    return (<p>Edit Line Page</p>)
-}
