+++ app/dashboard/lines/create/lib/create-line-action.ts
@@ -1,52 +1,9 @@
 'use server';
 
-import {revalidatePath, unstable_noStore as noStore} from "next/cache";
+import {revalidatePath} from "next/cache";
 import { redirect } from "next/navigation";
 import {z} from 'zod';
 import {createClient} from '@/utils/supabase/server';
-import {ITEMS_PER_PAGE} from "@/app/lib/utils";
-import {Agencies} from "@/app/lib/definitions";
-import FormData from "react";
-
-const supabase = createClient();
-
-export async function fetchLinePages(query: string) {
-    noStore();
-    
-    try {
-        let queryBuilder = supabase
-        .from("lines")
-        .select('*', { count: 'exact'})
-        
-        if (query) {
-            queryBuilder
-            .or(`line_number.ilike.%${query}%, legacy_line_number.ilike.%${query}%`)
-        }
-        const { count } = await queryBuilder
-        console.log('Pure count', count)
-        return Math.ceil(Number(count || 1) / ITEMS_PER_PAGE)
-    } catch (error) {
-        console.error('Database Error:', error)
-        throw new Error('Failed to fetch total number of lines.')
-    }
-}
-
-export async function fetchAgencies(): Promise<Agencies[]> {
-    noStore();
-    
-    try {
-        let queryBuilder = supabase
-        .from("agencies")
-        .select("id, created_at, name")
-        .order('id', { ascending: true })
-        const { data } = await queryBuilder
-        console.log('Agencies', data)
-        return data as Agencies[]
-    } catch (error) {
-        console.error('Database Error:', error)
-        throw new Error('Failed to fetch agencies')
-    }
-}
 
 const FormSchema = z.object({
     id: z.number(),
@@ -70,6 +27,9 @@ const FormSchema = z.object({
 const CreateLine = FormSchema.omit({ id: true, created_at: true, updated_at: true });
 
 export async function createLine(formData: FormData) {
+    
+    const supabase = createClient();
+    
     console.log('FormData', formData)
     const { line_number, legacy_line_number, units, agency_id, transport_type, line_type } = CreateLine.parse({
         line_number: formData.get('line_number'),
